[Driver] Pass JobAction to addClangTargetOptions.
In some cases we need additional job-specific info to figure out
the target-specific options. The patch adds an optional overload,
with the default implementation falling back to the existing variant
that only passes through the offloading kind.

We still need to maintain the old variant, because some compilation modes
call the function from the context where a JobAction is not available.